Added the print functionality
This closes #1
Using the "print" button now launch a print frame, which got enough
options in my opinion.
For now it print at the top (as a title) the document name & software
name, and the page number at the bottom. I'll change that later on to
give the user a way of editing it